backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
pymongo==4.7.2
motor==3.4.0
bcrypt==4.1.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pydantic==2.7.1

xgboost==2.0.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

transformers==4.40.2
torch==2.3.0
tokenizers==0.19.1

openai-whisper==20231117

tensorflow==2.16.1
opencv-python==4.9.0.80
Pillow==10.3.0

lxml==5.2.1